Latest Built

+++ mk-void-image.sh
@@ -4,10 +4,11 @@ set -e # Exit on first error
 set -x # Echo each command
 
 # Change to match latest values
-ARCH_IMAGE_VERSION="20181104"
-VOID_ROOTFS_VERSION="20181111"
+ARCH_IMAGE_VERSION="20190112"
+VOID_LOCAL_ROOTFS_VERSION="20190215"
+VOID_REMOTE_ROOTFS_VERSION="20181111"
 
-# We will use anarsoul's Arch Linux image to extract kernel + uboot
+# Use anarsoul's Arch Linux image to extract kernel + uboot
 ARCH_IMAGE_NAME="archlinux-xfce-pine64-${ARCH_IMAGE_VERSION}.img"
 ARCH_IMAGE_URL="https://github.com/anarsoul/linux-build/releases/download/${ARCH_IMAGE_VERSION}/${ARCH_IMAGE_NAME}.xz"
 
@@ -18,17 +19,29 @@ if [ ! -f "$ARCH_IMAGE_NAME" ]; then
   xz --decompress ${ARCH_IMAGE_NAME}.xz
 fi
 
-# We'll use aarch64 ROOTFS supplied by Void Linux
-VOID_ROOTFS_NAME="void-aarch64-ROOTFS-${VOID_ROOTFS_VERSION}.tar.xz"
-VOID_ROOTFS_URL="https://alpha.de.repo.voidlinux.org/live/current/${VOID_ROOTFS_NAME}"
+# Use locally built Platform/Root FS if available else fallback to void supplied Root FS.
+VOID_PLATFORMFS_NAME="void-pine64-A64-PLATFORMFS-${VOID_LOCAL_ROOTFS_VERSION}.tar.xz"
+VOID_LOCAL_ROOTFS_NAME="void-aarch64-ROOTFS-${VOID_LOCAL_ROOTFS_VERSION}.tar.xz"
+VOID_REMOTE_ROOTFS_NAME="void-aarch64-ROOTFS-${VOID_REMOTE_ROOTFS_VERSION}.tar.xz"
+VOID_ROOTFS_URL="https://alpha.de.repo.voidlinux.org/live/current/${VOID_REMOTE_ROOTFS_NAME}"
 
-if [ ! -f "$VOID_ROOTFS_NAME" ]; then
-  echo "Downloading Void Linux ROOTFS"
-  wget "$VOID_ROOTFS_URL" 
+if [ -f "$VOID_PLATFORMFS_NAME" ]; then
+  echo "Platform FS found. Using it over Root FS."
+  VOID_ROOTFS_NAME="$VOID_PLATFORMFS_NAME"
+elif [ -f "$VOID_LOCAL_ROOTFS_NAME" ]; then
+  echo "Root FS found. Using local copy."
+  VOID_ROOTFS_NAME="$VOID_LOCAL_ROOTFS_NAME"
+else
+    if [ ! -f "$VOID_REMOTE_ROOTFS_NAME" ]; then
+      echo "Downloading Void Linux Root FS from Void Servers."
+      wget "$VOID_ROOTFS_URL" 
+    fi
+    VOID_ROOTFS_NAME="$VOID_REMOTE_ROOTFS_NAME"
 fi
 
+
 # This will be our void image 
-IMAGE_NAME="${1:-void-aarch64-pine-A64plus.img}"
+IMAGE_NAME="${1:-void-pine64-A64plus.img}"
 BOOTLOADER="u-boot-sunxi-with-spl-pine64.bin"
 IMAGE_SIZE=4096M # 4 GB
 PART_POSITION=20480 # K
@@ -110,16 +123,18 @@ sleep 2
 sync
 
 # Setup /etc/fstab
-#BOOT_UUID=$(blkid --output=udev /dev/loop0p1 |grep _UUID= |cut -d= -f2) # Not Needed
+SWAP_UUID=$(blkid --output=udev /dev/loop0p2 |grep _UUID= |cut -d= -f2)
+echo 'UUID='${SWAP_UUID}'	swap	swap	defaults	0	0' >> sdcard/root/etc/fstab
 ROOTFS_UUID=$(blkid --output=udev /dev/loop0p3 |grep _UUID= |cut -d= -f2)
 echo 'UUID='${ROOTFS_UUID}'	/	ext4	defaults,rw,noatime	0	1' >> sdcard/root/etc/fstab
+#BOOT_UUID=$(blkid --output=udev /dev/loop0p1 |grep _UUID= |cut -d= -f2) # Not Needed
 #echo 'UUID='${BOOT_UUID}'	/boot	vfat	defaults,rw,noatime	0	0' >> sdcard/root/etc/fstab # Not Needed
 
 # Start some services by default
-ln -s /etc/sv/sshd sdcard/root/etc/runit/runsvdir/default/
-ln -s /etc/sv/ntpd sdcard/root/etc/runit/runsvdir/default/
-ln -s /etc/sv/dhcpcd sdcard/root/etc/runit/runsvdir/default/
-ln -s /etc/sv/agetty-ttyS0 sdcard/root/etc/runit/runsvdir/default/
+[[ -L "sdcard/root/etc/runit/runsvdir/default/sshd" ]] || ln -s /etc/sv/sshd sdcard/root/etc/runit/runsvdir/default/
+[[ -L "sdcard/root/etc/runit/runsvdir/default/ntpd" ]] || ln -s /etc/sv/ntpd sdcard/root/etc/runit/runsvdir/default/
+[[ -L "sdcard/root/etc/runit/runsvdir/default/dhcpcd" ]] || ln -s /etc/sv/dhcpcd sdcard/root/etc/runit/runsvdir/default/
+[[ -L "sdcard/root/etc/runit/runsvdir/default/agetty-ttyS0" ]] || ln -s /etc/sv/agetty-ttyS0 sdcard/root/etc/runit/runsvdir/default/
 
 
 # UBoot
